requirements.txt
aiogram>=3.18.0
aiohttp>=3.9.3
python-dotenv>=1.0.0
pandas>=2.2.3
openpyxl>=3.1.5
pytz>=2024.1
